Extract chat request handling core
